Switch keyboard scanning timer to single shot
Previously the timer used for scanning the keyboard matrix was setup to
be repeating but this caused problems when the main loop was executing
long running tasks when too many of these timer events would fill up
the application scheduling queue. The timer is now single shot and is
restarted each time the main loop finally gets around to running the
callback for the previous timer tick. This will add more jitter to the
sample times but that doesn't matter for keyboard scanning anyway.