Add wakeup on spacebar press
Added kbmatrixConfigureForWakeupOnSpacebar() function which configures
the row and column pins corresponding to the space key. The column to
which the space key is connected is pulled low and the row to which it
is attached is setup to pull the interrupt pin on the MCP23018 low
when the space key closes the circuit. The MCP23018 interrupt pin is
configured as a wakeup source for the nRF51 CPU.