Send back actual battery level
Now uses the ADC to sample the voltage of the battery instead of using
a simulator to fake values to send back via the BLE battery service.
It is quite cool that the ADC in the nRF51 can do this without using
any external pins/connections. Internally it is able to compare 1/3Vcc
to a 1.2VBG.

Other updates include:
* No longer debug dump the HID output report to the UART.
* Renamed the keyboard in the advertising data.
* Sample the battery every minute.