Version 3 of the PCB design
This is the design being sent off to OSHPark on May 6th, 2018.

This version of the PCB no longer uses the 20-pin Molex connector on
the bottom of the Dynastream N5150M5CD module and instead uses the
pads around the perimeter of the N5150M4CD module (actually the same
module but without the Molex and SWD connectors) instead. The module
will now mount on the bottom of the PCB with the module can sticking up
through a hole in this new PCB. This will give us extra height
clearance and correct the problem that the current PCB doesn't fit in
the original keyboard case.

I updated the firmware to accomodate the change in pin mappings for
UART, I2C, LEDs, etc.